# schema/openapi.yaml
...
openapi: 3.1.0
info:
  title: slim-m API
  version: 0.5.0
  description: >
    The server surface: liveness, auth with opaque rotating tokens, channel
    management, text messaging with reactions, polls, deletion, and full-text
    search, user profiles and the member list, devices, blocking, report
    intake and triage, invite codes, read state, bundled catch-up sync,
    account deletion, admin-issued password reset, role management and
    channel permission overwrites, presence and typing indicators, and the
    WebSocket event envelope.

    Identity and ordering are separate concerns. A message's identity is a
    client-generated UUIDv7, which also makes sending idempotent; its order is a
    per-channel monotonic `seq` assigned by the server, which is the
    authoritative cursor for pagination and sync.

    Authentication is a bearer access token. Access tokens are short-lived and
    exchanged at `/auth/refresh`, which rotates the refresh token. Opening a
    WebSocket uses a single-use ticket from `/auth/ws-ticket`, because an
    upgrade request cannot reliably carry an Authorization header.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://{host}
    variables:
      host:
        default: localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: The user's UUID.
        schema:
          type: string
          format: uuid
    get:
      operationId: getUser
      tags: [users]
      summary: A single public profile.
      description: >
        A deleted or anonymized account answers 404, exactly like an id that
        was never used, so this cannot confirm someone deleted their account.
      responses:
        "200":
          description: The user's public profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: List the channels the caller can view.
      responses:
        "200":
          description: The visible channels.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Create a channel.
      description: Requires the MANAGE_CHANNELS permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelRequest"
      responses:
        "200":
          description: The created channel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /channels/{channelId}:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
    patch:
      operationId: updateChannel
      tags: [channels]
      summary: Rename a channel and/or replace its topic.
      description: >
        Requires MANAGE_CHANNELS at the deployment level, the same check
        `createChannel` uses - the topic half of this route is gated
        identically, not a separate permission. Checked at the deployment
        level rather than per-channel so it stays well defined for a channel
        that is being (or has been) deleted; a caller without
        MANAGE_CHANNELS anywhere cannot tell a real channel id from a fake
        one either way.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChannelRequest"
      responses:
        "200":
          description: The updated channel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteChannel
      tags: [channels]
      summary: Soft-delete a channel.
      description: >
        Requires MANAGE_CHANNELS at the deployment level. Refuses to delete
        the deployment's last live channel, since a deployment with zero
        channels has nowhere for anyone to land. Deleting an
        already-deleted channel is not an error; a channel id that was
        never real is a 404.
      responses:
        "204":
          description: The channel was deleted (or was already gone).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /dms:
    get:
      operationId: listDirectMessages
      tags: [dms]
      summary: List the caller's direct-message conversations.
      description: >
        Each entry is a DM channel: its id (used with the ordinary
        `/channels/{channelId}/messages`, search, and `/sync` routes exactly
        like any other channel), the other participant's public profile, and
        the caller's unread count in it. Most recently active first.
      responses:
        "200":
          description: The caller's DM conversations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DmConversation"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /dms/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: The other participant's UUID.
        schema:
          type: string
          format: uuid
    post:
      operationId: openDirectMessage
      tags: [dms]
      summary: Open (or return) the DM channel with another user.
      description: >
        Idempotent and race-safe: opening the same pair twice, even
        concurrently, converges on one channel. Membership of the pair is the
        only thing that grants access to it; the deployment's role and
        channel-overwrite system, ADMINISTRATOR included, does not apply.
        Refused if either party has blocked the other (both directions
        checked), or if the target does not exist.
      responses:
        "200":
          description: The DM conversation, existing or newly opened.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DmConversation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An opaque access token from register, login, or refresh.
  parameters:
    ChannelId:
      name: channelId
      in: path
      required: true
      description: The channel's UUID.
      schema:
        type: string
        format: uuid
  responses:
...
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, expired, or revoked credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: >
        The caller lacks the required permission. A channel that does not exist
        is refused identically, so its existence is not observable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The resource does not exist within a scope the caller can see.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The request collided with existing state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: A short, stable, human-readable reason.
...
    UserProfile:
      type: object
      required: [id, username, display_name, created_at]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        created_at:
          type: integer
          format: int64
          description: Unix milliseconds.
        avatar_updated_at:
          type: [integer, "null"]
          format: int64
          description: >
            When this user's avatar was last set, or null for no avatar. Not
            itself fetchable: a client appends it as a cache-busting query
            parameter on `getAvatar`, which ignores the query string and
            serves whatever is currently stored. Absent on servers older
            than the avatar feature, which a client must treat as unknown
            (no cache key), not as "no avatar".
        roles:
          type: array
          items:
            type: string
          description: >
            Role names this member holds, for a client to render a badge
            (e.g. "Op") beside them. Excludes `@everyone`: every member holds
            that one, so including it would put a meaningless badge on every
            row. Empty rather than absent for a member with nothing beyond
            it. Deliberately no colour field: badges use the design system's
            accent, not a per-role one. Absent on servers older than this
            field, which a client must treat as unknown rather than "no
            roles".
...
    Channel:
      type: object
      required: [id, name, kind, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        kind:
          type: string
          enum: [text, voice]
        topic:
          type: [string, "null"]
          maxLength: 256
          description: >
            A one-line description shown beside the name in the client's
            channel header, or null for no topic. Absent on servers older
            than this field, which a client must treat as unknown rather
            than "no topic".
        created_at:
          type: integer
          format: int64
          description: Unix milliseconds.
    CreateChannelRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        kind:
          type: string
          enum: [text, voice]
          default: text
    UpdateChannelRequest:
      type: object
      description: >
        At least one of `name` and `topic` must be present; a request with
        neither is a 400. Each is independent: sending only `topic` leaves
        the name untouched, and vice versa.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        topic:
          type: string
          maxLength: 256
          description: >
            A blank or whitespace-only value clears the topic back to null
            rather than storing an empty string.
    DmConversation:
      type: object
      required: [channel_id, user, unread, created_at]
      properties:
        channel_id:
          type: string
          format: uuid
        user:
          $ref: "#/components/schemas/UserProfile"
        unread:
          type: integer
          format: int64
        created_at:
          type: integer
          format: int64
          description: Unix milliseconds the DM channel was first opened.
...